Initial implementation of object cache.
Grace classes given on the command line are serialized and written to /object_cache.
When executed again the cache is checked and the object is de-serialized and executed.
On any sort of cache-related error, or if the --no-cache option ise given the previous
behavior of reading, parsing, translating, then executing is used.